requirements.txt
AnisoCADO==0.3.0
anyio==4.7.0
astar-utils==0.3.1
astropy==6.1.7
astropy-iers-data==0.2024.12.23.0.33.24
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docutils==0.19
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
more-itertools==10.5.0
numpy==1.26.4
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
Pyckles==0.4.0
pyerfa==2.0.1.5
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
result==0.17.0
scipy==1.14.1
ScopeSim==0.9.1
ScopeSim_Templates==0.6.1
six==1.17.0
skycalc-ipy==0.5.2
sniffio==1.3.1
soupsieve==2.6
speXtra==0.41.3
synphot==1.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
